+++ audionimbus/src/effect/pathing.rs
@@ -47,7 +47,7 @@ use crate::simulation::{SimulationOutputs, Simulator, Source};
 ///     .with_direct(DirectSimulationSettings {
 ///         max_num_occlusion_samples: 4,
 ///     })
-///    .with_reflections(ReflectionsSimulationSettings::Convolution {
+///     .with_reflections(ReflectionsSimulationSettings::Convolution {
 ///         max_num_rays: 4096,
 ///         num_diffuse_samples: 32,
 ///         max_duration: 2.0,